Removes pregameState and changes endGame format
Removes the pregameState state. It was used to prevent the creator from
knowing if they started as white or black by preventing it from
displaying differently before the game started, but it now has a default
way it renders if the game hasnt started. Also changes the endGame state
format to be more readable, and adds a reason to the win/loss screen.